changing to interface /cc @drsnyder
