+++ lib/Wikia/tests/Service/User/PreferenceServiceImplTest.php
@@ -8,7 +8,7 @@ use PHPUnit_Framework_TestCase;
 use Wikia\Domain\User\Preferences\UserPreferences;
 use Wikia\Persistence\User\Preferences\PreferencePersistence;
 
-class PreferenceServiceTest extends PHPUnit_Framework_TestCase {
+class PreferenceServiceImplTest extends PHPUnit_Framework_TestCase {
 	const TEST_WIKI_ID = 123;
 
 	/** @var int */
@@ -47,7 +47,7 @@ class PreferenceServiceTest extends PHPUnit_Framework_TestCase {
 
 	public function testGetFromDefault() {
 		$defaultPrefs = ["pref1" => "val1"];
-		$preferences = new PreferenceService($this->cache, $this->persistence, [], $defaultPrefs, []);
+		$preferences = new PreferenceServiceImpl($this->cache, $this->persistence, [], $defaultPrefs, []);
 
 		$this->assertEquals("val1", $preferences->getFromDefault("pref1"));
 		$this->assertNull($preferences->getFromDefault("pref2"));
@@ -55,7 +55,7 @@ class PreferenceServiceTest extends PHPUnit_Framework_TestCase {
 
 	public function testGet() {
 		$this->setupServiceExpects();
-		$preferences = new PreferenceService($this->cache, $this->persistence, [], [], []);
+		$preferences = new PreferenceServiceImpl($this->cache, $this->persistence, [], [], []);
 
 		$this->assertEquals("en", $preferences->getGlobalPreference($this->userId, "language"));
 		$this->assertEquals("1", $preferences->getGlobalPreference($this->userId, "marketingallowed"));
@@ -69,14 +69,14 @@ class PreferenceServiceTest extends PHPUnit_Framework_TestCase {
 
 	public function testGetWithHiddenNoDefaults() {
 		$this->setupServiceExpects();
-		$preferences = new PreferenceService($this->cache, $this->persistence, ["marketingallowed"], [], []);
+		$preferences = new PreferenceServiceImpl($this->cache, $this->persistence, ["marketingallowed"], [], []);
 		$this->assertEquals("1", $preferences->getGlobalPreference($this->userId, "marketingallowed", null, true));
 		$this->assertNull($preferences->getGlobalPreference($this->userId, "marketingallowed"));
 	}
 
 	public function testGetWithHiddenAndDefaults() {
 		$this->setupServiceExpects();
-		$preferences = new PreferenceService($this->cache, $this->persistence, ["marketingallowed"], ["marketingallowed" => "0"], []);
+		$preferences = new PreferenceServiceImpl($this->cache, $this->persistence, ["marketingallowed"], ["marketingallowed" => "0"], []);
 		$this->assertEquals("1", $preferences->getGlobalPreference($this->userId, "marketingallowed", null, true));
 		$this->assertEquals("0", $preferences->getGlobalPreference($this->userId, "marketingallowed"));
 		$this->assertNull($preferences->getGlobalPreference($this->userId, "unsetpreference"));
@@ -84,21 +84,21 @@ class PreferenceServiceTest extends PHPUnit_Framework_TestCase {
 	
 	public function testSet() {
 		$this->setupServiceExpects();
-		$preferences = new PreferenceService($this->cache, $this->persistence, [], [], []);
+		$preferences = new PreferenceServiceImpl($this->cache, $this->persistence, [], [], []);
 		$preferences->setGlobalPreference($this->userId, "newpreference", "foo");
 		$this->assertEquals("foo", $preferences->getGlobalPreference($this->userId, "newpreference"));
 	}
 
 	public function testSetNullWithDefault() {
 		$this->setupServiceExpects();
-		$preferences = new PreferenceService($this->cache, $this->persistence, [], ["newpreference" => "foo"], []);
+		$preferences = new PreferenceServiceImpl($this->cache, $this->persistence, [], ["newpreference" => "foo"], []);
 		$preferences->setGlobalPreference($this->userId, "newpreference", null);
 		$this->assertEquals("foo", $preferences->getPreferences($this->userId)->getGlobalPreference("newpreference"));
 	}
 
 	public function testSetNullWithoutDefault() {
 		$this->setupServiceExpects();
-		$preferences = new PreferenceService($this->cache, $this->persistence, [], [], []);
+		$preferences = new PreferenceServiceImpl($this->cache, $this->persistence, [], [], []);
 		$preferences->setGlobalPreference($this->userId, "newpreference", null);
 		$this->assertNull($preferences->getPreferences($this->userId)->getGlobalPreference("newpreference"));
 	}
